Write a program that multiplies two positive numbers
num1 and num2 will be passed in base 10

If the number of arguments is incorrect, print Error, followed by a new line, and exit with a status of 98.

num1 and num2 should only be composed of digits. If not, print Error, followed by a new line, and exit with a status of 98