Write a function that inserts a new node at a given position
idx is the index of the list where the new node should be added. Index starts at 0. Return the address of the new node , or NULL if it failed. If it is not possible to add the new node at index idx, do not add the new node and return NULL